Start refactor by split code into files
