--- main.py
@@ -1,5 +1,0 @@
-"""Play Wordle running this module."""
-from src.game import Wordle
-
-wordle = Wordle()
-wordle.game()
